Added first integration test
Added a bash integration test to the test suite. This sends a successful ping to our server with an empty body and is returned a successful response.
This test uses curl and port 80.

Allows us to test the entire process of running the server, sending a command, and receiving a response. Our unit tests tests our functions and code components, so this test
is an additional layer that tests all the moving parts together. It is more general than a unit test.

Used Bash and created a shell script; added a test rule into our CMAKE to run this integration test with `make test`.
The integration test compares the output of the server response to a file
with the standard expected response.

Change-Id: I7262a61639214cd63709cccfcd349ea7996316db