Added first integration test
# What
Added a bash integration test to the test suite

# Why
Allows us to test the entire process of running the server, sending a command, and receiving a response.

# How
Used Bash and created a shell script; added a test rule into our CMAKE to run this integration test with `make test`. The integration test compares the output of the server response to a file with the standard expected response.

Change-Id: I7262a61639214cd63709cccfcd349ea7996316db