fixed directory bug on integration test
The integration test was failing since the command to start the server `./bin/server ../conf/sample.conf &` was directing to the wrong path where it could not find the executable. I updated it to `./build/bin/server ./conf/sample.conf &` since the CMAKE file defines the directory as the project source directory. the new pathway will direct it to the correct executable.

Change-Id: I0b910cecfe469d061e5bacc02d8b6eb59ea15969