Added a logger
#What
Configured boost::log to be used for web server logging. Log files are stored in the logs/ folder. Each line contains a timestamp, thread ID, and message. Log files reset at midnight and when they reach a size of 10 MB.

#Why
Logging is useful for debugging issues and generally monitoring the health of a web server.

#How
https://www.boost.org/doc/libs/1_74_0/libs/log/doc/html/index.html
https://www.boost.org/doc/libs/1_74_0/libs/log/doc/html/log/reference.html
https://www.boost.org/doc/libs/1_74_0/libs/log/doc/html/boost/log/add_file_log.html
https://boost-log.sourceforge.net/libs/log/doc/html/boost/log/add_console_lo_idp23239792.html

#Testing
No tests yet.

Change-Id: I4a53acfb155c9f6ce7092a7a3f6d43db49b89253