Added severity logging and basic error checking
#What
Logs now contain a severity level before the message. They also now record errors with config file parsing and server startup.

#Why
It is important to be able to distinguish what information in the log is urgent/important, and it is useful to log where the server failed.

#How
Added an enum defining different log entry types and a map to what is outputted in the log file for each of them. Replaced previous cerr messages with BOOST_LOG_TRIVIAL() function calls.

#Testing
No tests yet. Will need to create some unit tests that check that the log successfully records invalid config files and an integration test that checks for the creation of a log file.

Change-Id: I27b5102b2d5ca64cddcddd398b4ecb2ec0d3e1a9