Added some logging integration tests
#What
Added integration tests that test the following:
1. Log file is created
2. Log file logs invalid port number
3. Log file logs invalid config file

#Why
It is important to make sure that the log file works and that fatal errors show up in the file.

#How
Created shell scripts that use the test -f command to check the existence of a file and the grep command to check for the existence of a string containing the fatal error within the file.

#Testing
This change contains them lol.

Change-Id: I6cde2a56edf68745edc972cf4dcbc7cf78582977