Created class to manage echo handler
#What
Maintains the `echo` functionality but moved the logic into a request handler class.

#Why
We are going to create a static file request handler, so this will match the structure of that handler

#How
Created new handler.cc and handler.h files as well as edited CMAKE

#Testing
The current test suite passes, which means our functionality holds up correctly.

Change-Id: I337c5afa22d9d7012337ac37a45defce71c9c291