Added logging for terminations, requests, and config parsing
#What
Made sure logging worked in all files that required logging.

#Why
We want to be able to log incoming requests so we can verify activity in the console as it happens to the server, and log info about config parsing.

#How
Added correct include header files and used BOOST_LOG_TRIVIAL(...) to log desired messages.

#Testing
None.

Change-Id: I097e80fb82cfb4fa6f584e24d54c0ce7a2e8b636