Added logic for static file parsing and serving
#What
This adds functionality to serve static files to our server under the pathway /static. This also maintains echo functionality under the pathway /echo (for the sample config). All other paths serve a 404 error. the pathway for echo and static are defined in the config file

#Why
This adds functionality for serving static files. It adds more configurability to the server

#How
Created new handler.cc and handler.h files as well as edited CMAKE for the static class (similar to how we had done for the echo class). I also added to the config parser class to create functions that look for the static and echo keywords in a config file.

#Testing
The current test suite passes; I am going to add more unit tests in a follow up CL.

to run:
`./build/bin/server ./conf/sample.conf` in one terminal from the project directory, and use curl (for ex `curl localhost:80/echo`) in another terminal

Change-Id: Ie3ba97f2b9b3b503f9afea0d52bfeb47e03fdce8