Added stock replies to handle 200, 400, and 404 statuses
#What
Created function to handle the usual format for 200 OK, 400 Bad Request, and 404 Not Found statuses.

#Why
We want to be able to easily format these statuses in our session.cc and static_handler.cc files later on when a request is sent and we want to reply with the correct status.

#How
Added new functions and namespace definitions for 200 OK, 400 Bad Request, and 404 Not Found statuses.

Change-Id: I90291d34c33c13c4bcab77d9e2fb8b97059154ad