Added static integration test and fixed echo integration test
#What
Added a new integration test for static functionality that verifies that the server returns the contents of static/index.html. Fixed an existing integration test for echo functionality that previously verified that the server returns a 404 not found when it should actually verify that the server just sends the request back. Also renamed some log file integration tests to be more consistent with naming convention.

#Why
It is important to make sure that the various components of our web server work together properly to handle static and echo requests correctly. It is also important for our naming convention to be consistent across our code base because we will be handing our code over to other teams in later assignments, and we want them to be able to understand it easily.

#How
For the static integration test, wrote a new shell script that spins up an instance of the web server, sends a request to /static/index.html, and compares the output to what it should be. For the echo integration test, modified the existing shell script that spins up an instance of the web server, sends a request to /echo, and makes sure that the output is the same as what was sent.

#Testing
This whole change is all about testing lol.

Change-Id: Ia520053c0d5da14af37fd398bf0042254f6eb7e7