Fixed a bug where 404s were being logged as good requests
#What
Previously, requests to the web server that did not access an echo or static endpoint (error 404s) were being logged as good requests. They are now logged as 404s.

#Why
It is useful to have a log that contains 404s when debugging issues with echo and static endpoints.

#How
Moved logging statements for good requests inside if statements that check that the request is headed to an echo or static endpoint. Added logging statements for requests that are not headed to an echo or static endpoint.

#Testing
Will add an integration test for 404s in a future change.

Change-Id: I7cb67a1ba57ea5c0bc91ef95a69aea265bbfac78