Added an integration test that verifies that requests to an invalid URI return an error 404
#What
Added an integration test that makes a curl request to the /asdf endpoint (which does not exist) and verifies that it returns an error 404.

#Why
Endpoints that do not serve static content or implement echo functionality should return a 404 not found.

#How
Wrote a shell script that spins up an instance of the web server and checks the output of a curl HTTP request to /asdf to make sure it matches what is expected.

#Testing
There are two ways to run this test:
1. As part of the full test suite with `make test`
2. Manually by executing `./tests/successful_get_404_response_integration_test.sh`
If you are using the second method, then you may need to first execute `chmod u+x ./tests/successful_get_404_response_integration_test.sh`

Change-Id: I2001cae7455f96210e92b4894e5c6b0972534b01