Code cleanup and additional logging
#What
General code hygiene (curly braces should all be on a new line, two spaces for indents, etc.), clearer comments, and more logging.

#Why
We will be handing our code off to another team in a later assignment and we want them to be able to read and understand it easily. Extra logging is useful for debugging.

#How
Rearranged some code, added some logging statements.

This change does not make any changes to the functionality of the server.

Change-Id: Ib97f3feb155a2708d2a2551126d2647afb193561