Code cleanup, commenting, and logging
#What
Clearer comments and more logging.

#Why
We will be handing our code off to another team in a later assignment and we want them to be able to read and understand it easily. Extra logging is useful for debugging.

#How
Rearranged some code, added some logging statements.

This change does not make any changes to the functionality of the server.

Change-Id: I68350aacd067f9d8e9548760f07b39652137f78c