Updated docker/cloudbuild.yaml
#What
A previous commit assumed that the "name" fields in cloudbuild.yaml are used for naming the various steps, when in fact they are used to specify a cloud builder. Their renaming caused our cloud build to break, so they are being changed back.

#Why
A working cloud build is necessary for successful deployment to GCE.

#How
Changed the name fields back to "gcr.io/cloud-builders/docker" (a Docker cloud builder). Added comments naming the various steps instead.

#Testing
The only way to test is to merge into main and trigger a cloud build in Google Cloud.

Change-Id: I1b9e5d9a15b0e61d446336ce3ca90531e0503c31