Serves static files on relative path based on classwide API
What: the class decided on a format for config files, so this updates how the static handler is parsed. it handles both absolute and relative pathways

Why: the classwide api instructs configs to be built a certain way that was different than our previous implementation. i changed the config parser to reflect the new config format

How: i use a map to map the uri pathway to the resource file location, using that mapping we either serve a static file or a 404 error if not found.

Test: current test suite passes, meaning functionality hasn't changed. the test configs have been changed to match the classwide API

Change-Id: I3661728dcfef06d7f8871306b753c3f081734f89