Skeleton files & configuration for CrudHandler
#Why
Setup needed for CrudHandler

#How
Created .h and .cc files for CrudHandler with basic class definition.  In config file, CrudHandler is configured to listen for requests at /api & performs crud operations on files in data path which is set to /mnt/crud. Created mnt/crud directory that will serve as the filesystem that CrudHandler will store and retrieve data/files from.

Change-Id: I87672a930c695bd3eeffc43d1079f7d65f678aa3