Added some more machine-parsable log output
#What
Added more logs that would be machine-parsable.

#Why
We want to output a log line that can be easily searched for / match among all your other log lines and that contains metrics that can be matched and extracted with a regular expression.

#How
Added log lines that can be easily searched.

#Testing
None.

Change-Id: I94537214b6fa9773e799083cf2180cdbbe08a00b