Created a new class for the markdown handler
#What
Created a new MarkdownHandler class that is a subclass of the RequestHandler superclass. The handle_request function currently just returns a 200 OK response similar to the health and sleep handlers. Proper functionality will be implemented in a later change.

#Why
Another team member will be implementing config parsing and factory creation for the markdown handler in a separate change, so it is useful to have a request handler that can be instantiated.

#How
Created a new class that extends the request handler class and implements the handle_read function.

Change-Id: I16eb525c82157693296da538b869d1108bfcfa43