Started implementing markdown handler
#What
Started implementing the markdown handler with basic functionality for converting from Markdown to HTML format including:
- Small, medium, and large headings
- Regular text
- Code blocks

#Why
These are the most commonly used features of Markdown.

#How
MarkdownHandler::handle_request method reads through Markdown document and converts "# [text]" to "<h1>[text]</h1>", "## [text]" to "<h2>[text]</h2>", and "### [text]" to "<h3>[text]</h3>". Regular "[text]" is converted to "<p>[text]</p>" and "```" is converted to either "<code>" or "</code>" depending on whether or not we are already inside a code block.

#Bugs
There is currently no support for in-line code, links, bold text, italicized text, underlined text, bulleted lists, and numbered lists. Some (if not all) of these will be added in a later change.

#Testing
No tests yet. Will be added in a later change.

Change-Id: I1bca5eb514d5a4b30e652d8fc6f67a97780bbf0a