Renamed markdown_handler to markdown_request_handler for consistency and added garbage cleanup
#What
Renamed markdown_handler.cc and markdown_handler.h to markdown_request_handler.cc and markdown_request_handler.h for consistency. Also added garbage cleanup to Markdown handler; it now deletes the HTML file that it generates right after it copies it into the request response.

#Why
We want all of our request handlers to be consistent in their naming. And we don't want to clutter our project directory with a bunch of HTML files.

#How
Renamed the files and made the proper adjustments in CMakeLists.txt and markdown_handler_factory.cc so that the code compiles and runs as it did before. Added a line to the Markdown handler that deletes the file after reading it into the request response.

Change-Id: Icdccd32bd97021b0c5bc7790362fc459d9900993